# api/docs/openapi.yaml
openapi: 3.0.3
info:
  title: AutoLlama v2.2 - Intelligent Context API
  description: |
    **AutoLlama v2.2 "Intelligent Context Llama"** - Enhanced RAG platform with intelligent contextual retrieval and advanced document understanding.
    
    ## 🚀 Enhanced Features v2.2
    - **35-60% Better Retrieval** through intelligent contextual processing
    - **Document Intelligence**: Automatic type detection and structural analysis
    - **Semantic Chunking**: Boundary-aware segmentation with hierarchical understanding
    - **Advanced Context Generation**: Document-aware prompts with retry logic
    - **Enhanced Metadata**: 11 new contextual fields for performance analysis
    - **Production-Ready**: Ultra-safe processing with comprehensive error handling
    
    ## Processing Capabilities
    - **Multi-format Processing**: PDF, DOCX, EPUB, CSV, URL fetching
    - **Hybrid Search**: Vector similarity + BM25 text search  
    - **Real-time Processing**: SSE streaming with progress updates
    - **Intelligent Chunking**: Semantic, structural, and hierarchical strategies
    - **Contextual Analysis**: Sentiment, topics, emotions, categories + document structure
    
    ## Authentication
    This API is designed for secure homelab deployment behind Tailscale VPN.
    
    ## Rate Limiting & Error Handling
    - **Search endpoints**: 100 requests/minute
    - **Processing endpoints**: 10 requests/5 minutes  
    - **Health endpoints**: No limits
...
  version: 2.2.0
...
  contact:
...
servers:
...
paths:
...
  /search:
    get:
      tags: [Search & Retrieval]
      summary: Hybrid search
      description: |
        Perform hybrid search combining vector similarity and BM25 text search.
        Returns ranked results with similarity scores and contextual information.
      parameters:
        - name: q
          in: query
          required: true
          description: "🔍 Search query text - Try: 'machine learning', 'python tutorial', or any topic"
          schema:
            type: string
            minLength: 1
            maxLength: 1000
          example: "machine learning tutorial"
        - name: limit
          in: query
          description: Maximum number of results to return
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
        - name: offset
          in: query
          description: Number of results to skip for pagination
          schema:
            type: integer
            minimum: 0
            default: 0
        - name: includeChunks
          in: query
          description: Include related chunks in results
          schema:
            type: boolean
            default: false
        - name: threshold
          in: query
          description: Minimum similarity threshold (0.0-1.0)
          schema:
            type: number
            minimum: 0.0
            maximum: 1.0
            default: 0.7
      responses:
        '200':
          description: Search results with similarity scores
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SearchResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '500':
          $ref: '#/components/responses/InternalError'
...
  /search/documents:
    get:
      tags: [Search & Retrieval]
      summary: Document search
      description: Search documents by title, content, and metadata
      parameters:
        - name: q
          in: query
          required: true
          description: Search query
          schema:
            type: string
        - name: sortBy
          in: query
          description: Sort field
          schema:
            type: string
            enum: [created_at, updated_at, title, relevance]
            default: relevance
        - name: sortOrder
          in: query
          description: Sort order
          schema:
            type: string
            enum: [asc, desc]
            default: desc
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
      responses:
        '200':
          description: Document search results
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DocumentSearchResponse'
...
  /documents:
    get:
      tags: [Document Management]
      summary: List documents
      description: Get paginated list of documents with optional filtering
      parameters:
        - name: page
          in: query
          description: Page number (1-based)
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: limit
          in: query
          description: Items per page
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
        - name: search
          in: query
          description: Search term for filtering
          schema:
            type: string
        - name: sortBy
          in: query
          description: Sort field
          schema:
            type: string
            enum: [created_at, updated_at, title, chunk_count]
            default: created_at
        - name: sortOrder
          in: query
          description: Sort order
          schema:
            type: string
            enum: [asc, desc]
            default: desc
      responses:
        '200':
          description: Paginated document list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DocumentListResponse'
...
  /process-url:
    post:
      tags: [Content Processing]
      summary: Process URL synchronously
      description: |
        📄 Process content from a URL and return when complete.
        
        **Supported Content:**
        - Articles, blog posts, documentation
        - PDF documents, research papers  
        - Wikipedia pages, news articles
        
        **Try these example URLs:**
        - https://docs.python.org/3/tutorial/
        - https://en.wikipedia.org/wiki/Machine_learning
        - Any public webpage or PDF URL
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/URLProcessRequest'
            example:
              url: "https://docs.python.org/3/tutorial/introduction.html"
              chunkSize: 1000
              overlap: 100
              enableContextualEmbeddings: true
      responses:
        '200':
          description: URL processed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProcessingResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '500':
          $ref: '#/components/responses/ProcessingError'
  /process-url-stream:
    post:
      tags: [Content Processing]
      summary: Process URL with streaming
      description: |
        Process URL content with real-time progress updates via Server-Sent Events.
        Returns immediately with session ID for progress tracking.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/URLProcessRequest'
      responses:
        '200':
          description: Processing started, use SSE for progress
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StreamingResponse'
  /process-file:
    post:
      tags: [Content Processing]
      summary: Process file synchronously
      description: |
        Upload and process a file. Supports PDF, DOCX, EPUB, CSV, and text files.
        Maximum file size: 100MB.
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/FileProcessRequest'
            encoding:
              file:
                contentType: application/pdf, application/vnd.openxmlformats-officedocument.wordprocessingml.document, application/epub+zip, text/csv, text/plain
      responses:
        '200':
          description: File processed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProcessingResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '413':
          description: File size exceeds limit
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /process-file-stream:
    post:
      tags: [Content Processing]
      summary: Process file with streaming
      description: Process file with real-time progress updates via SSE
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/FileProcessRequest'
      responses:
        '200':
          description: Processing started
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StreamingResponse'
  /pre-upload-check:
    post:
      tags: [Content Processing]
      summary: Pre-upload validation
      description: |
        Validate system readiness before starting upload.
        Checks for stuck sessions, resource availability, and system health.
      responses:
        '200':
          description: System ready for upload
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PreUploadResponse'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'

  # Session Management
...
  /settings:
    get:
      tags: [Settings]
      summary: Get all settings
      description: Retrieve current system configuration
      responses:
        '200':
          description: Current settings
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SettingsResponse'

    post:
      tags: [Settings]
      summary: Update settings
      description: Update system configuration
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SettingsUpdateRequest'
      responses:
        '200':
          description: Settings updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SettingsResponse'
  /settings/{key}:
    get:
      tags: [Settings]
      summary: Get specific setting
      description: Retrieve a specific configuration value
      parameters:
        - name: key
          in: path
          required: true
          description: Setting key
          schema:
            type: string
      responses:
        '200':
          description: Setting value
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SettingResponse'
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: "⚠️ NOT IMPLEMENTED - Leave empty for local development"
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: "⚠️ NOT IMPLEMENTED - Leave empty for local development"
  responses:
    ValidationError:
      description: Validation error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error:
              message: "Validation failed"
              code: "VALIDATION_ERROR"
              details:
                - field: "chunkSize"
                  message: "Must be between 100 and 5000"
            statusCode: 400
...
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error:
              message: "Internal server error"
              code: "INTERNAL_ERROR"
              timestamp: "2024-01-15T10:30:00.000Z"
            statusCode: 500
    ServiceUnavailable:
      description: Service temporarily unavailable
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error:
              message: "Vector service not available"
              code: "SERVICE_UNAVAILABLE"
              timestamp: "2024-01-15T10:30:00.000Z"
            statusCode: 503
    ProcessingError:
      description: Processing failed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error:
              message: "Failed to process content"
              code: "PROCESSING_ERROR"
              details: "Unable to parse PDF content"
            statusCode: 500
  schemas:
...
    SearchResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        query:
          type: string
          example: "artificial intelligence"
        results:
          type: array
          items:
            $ref: '#/components/schemas/SearchResult'
        total:
          type: integer
          description: Total number of results
        limit:
          type: integer
        offset:
          type: integer
        searchType:
          type: string
          enum: [hybrid, vector, text]
          example: "hybrid"
        processingTime:
          type: number
          description: Search processing time in milliseconds
    SearchResult:
      type: object
      properties:
        chunk_id:
          type: string
          format: uuid
        chunk_text:
          type: string
        title:
          type: string
        url:
          type: string
          format: uri
        chunk_index:
          type: integer
        similarity_score:
          type: number
          minimum: 0
          maximum: 1
        sentiment:
          type: string
          enum: [positive, negative, neutral]
        contextual_summary:
          type: string
        main_topics:
          type: array
          items:
            type: string
        category:
          type: string
        chunks:
          type: array
          items:
            $ref: '#/components/schemas/ChunkData'
          description: Related chunks (if includeChunks=true)
...
    DocumentSearchResponse:
      type: object
      properties:
        success:
          type: boolean
        query:
          type: string
        documents:
          type: array
          items:
            $ref: '#/components/schemas/DocumentSummary'
        total:
          type: integer
        searchType:
          type: string
          example: "document"
...
    DocumentListResponse:
      type: object
      properties:
        success:
          type: boolean
        documents:
          type: array
          items:
            $ref: '#/components/schemas/DocumentSummary'
        pagination:
          $ref: '#/components/schemas/Pagination'
    DocumentSummary:
      type: object
      properties:
        document_id:
          type: string
          format: uuid
        title:
          type: string
        url:
          type: string
          format: uri
          nullable: true
        chunk_count:
          type: integer
        status:
          type: string
          enum: [processing, completed, failed]
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        metadata:
          type: object
...
    URLProcessRequest:
      type: object
      required: [url]
      properties:
        url:
          type: string
          format: uri
          example: "https://docs.python.org/3/tutorial/introduction.html"
          description: "🔗 Any public URL - Try Wikipedia, documentation sites, or blog posts"
        chunkSize:
          type: integer
          minimum: 100
          maximum: 5000
          default: 1000
          description: "📝 Text chunk size (1000 = ~2-3 paragraphs, good default)"
        overlap:
          type: integer
          minimum: 0
          maximum: 500
          default: 100
          description: "🔄 Character overlap between chunks (100 = ~1 sentence, prevents context loss)"
        enableContextualEmbeddings:
          type: boolean
          default: true
          description: "🧠 Generate smarter embeddings with context (recommended: true)"
    FileProcessRequest:
      type: object
      required: [file]
      properties:
        file:
          type: string
          format: binary
          description: File to process (PDF, DOCX, EPUB, CSV, TXT)
        chunkSize:
          type: integer
          minimum: 100
          maximum: 5000
          default: 1000
        overlap:
          type: integer
          minimum: 0
          maximum: 500
          default: 100
        enableContextualEmbeddings:
          type: boolean
          default: true
...
    ProcessingResponse:
      type: object
      properties:
        success:
          type: boolean
        sessionId:
          type: string
          format: uuid
        message:
          type: string
        results:
          type: object
          properties:
            documentsCreated:
              type: integer
            chunksProcessed:
              type: integer
            totalTokensUsed:
              type: integer
            processingTime:
              type: number
    StreamingResponse:
      type: object
      properties:
        success:
          type: boolean
        sessionId:
          type: string
          format: uuid
        message:
          type: string
        sseEndpoint:
          type: string
          description: Server-Sent Events endpoint for progress updates
...
    PreUploadResponse:
      type: object
      properties:
        success:
          type: boolean
        ready:
          type: boolean
        checks:
          type: object
          properties:
            systemHealth:
              type: boolean
            stuckSessions:
              type: boolean
            resourceAvailability:
              type: boolean
        recommendations:
          type: array
          items:
            type: string

    # Session Schemas
...
    SettingsResponse:
      type: object
      properties:
        success:
          type: boolean
        settings:
          type: object
          properties:
            processing:
              type: object
              properties:
                defaultChunkSize:
                  type: integer
                defaultOverlap:
                  type: integer
                enableContextualEmbeddings:
                  type: boolean
                maxFileSize:
                  type: integer
            search:
              type: object
              properties:
                defaultLimit:
                  type: integer
                maxLimit:
                  type: integer
                defaultThreshold:
                  type: number
            performance:
              type: object
              properties:
                requestTimeout:
                  type: integer
                maxConcurrentProcessing:
                  type: integer
    SettingsUpdateRequest:
      type: object
      properties:
        processing:
          type: object
          properties:
            defaultChunkSize:
              type: integer
              minimum: 100
              maximum: 5000
            defaultOverlap:
              type: integer
              minimum: 0
              maximum: 500
            enableContextualEmbeddings:
              type: boolean
        search:
          type: object
          properties:
            defaultLimit:
              type: integer
              minimum: 1
              maximum: 100
            defaultThreshold:
              type: number
              minimum: 0.0
              maximum: 1.0
    SettingResponse:
      type: object
      properties:
        success:
          type: boolean
        key:
          type: string
        value:
          oneOf:
            - type: string
            - type: number
            - type: boolean
            - type: object

    # Common Schemas
    Pagination:
      type: object
      properties:
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
        totalPages:
          type: integer
        hasNext:
          type: boolean
        hasPrev:
          type: boolean
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: object
          properties:
            message:
              type: string
            code:
              type: string
            timestamp:
              type: string
              format: date-time
            requestId:
              type: string
              nullable: true
            details:
              oneOf:
                - type: string
                - type: object
                - type: array
              nullable: true
        statusCode:
          type: integer

# No global security required - this is a local development API